Provide a GET Example with URL Parameter
Provide an example with URL Parameter.  Test the parameter with a sqs ReceiptHandle sample.  The ReceiptHandle, however, contains '/' and this will cause an issue with the http server.  The only way is to replace the '/' with '_' (or some other character) during the request and use strings.Replace() to replace it back.